third_party/blink/web_tests/external/wpt/referrer-policy/generic/inheritance/iframe-inheritance-javascript-child-expected.txt

This is a testharness.js-based test.
FAIL Referrer Policy: iframes with javascript url reuse referrer policy assert_equals: expected "http://web-platform.test:8001/" but got "http://web-platform.test:8001/custom"
FAIL Referrer Policy: iframes with javascript url reuse referrer policy 1 assert_equals: expected (undefined) undefined but got (string) "http://web-platform.test:8001/custom"
PASS Referrer Policy: iframes with javascript url reuse referrer policy 2
FAIL Referrer Policy: iframes with javascript url reuse referrer policy 3 assert_equals: expected (undefined) undefined but got (string) "http://web-platform.test:8001/"
Harness: the test ran to completion.
